move instrumentation to webapp/app.py
